+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "CGS minifier".
 - Tests for "Smooth factory".
 - Tests for "Sum of divisors".
+- Tests for "Numeral system".
 
 ## [1.15.0] - 2022-11-30
 ### Added